resolved the responsive issue, created common helper file
