+++ inc/customizer/react/src/responsive-color/control.js
@@ -1,4 +1,5 @@
 import ResponsiveColorComponent from './responsive-color-component';
+import { getResponsiveColorJs } from '../common/responsive-helper';
 
 export const responsiveColorControl = wp.customize.astraControl.extend( {
 	renderContent: function renderContent() {
@@ -10,41 +11,7 @@ export const responsiveColorControl = wp.customize.astraControl.extend( {
 	},
 	ready: function() {
 
-		'use strict';
-
-		jQuery('html').addClass('responsive-background-color-ready');
-
-		let device = jQuery('.wp-full-overlay-footer .devices button.active').attr('data-device')
-
-		jQuery( '.customize-control-ast-responsive-color .customize-control-content .ast-color-picker-alpha' ).removeClass( 'active' );
-
-		jQuery( '.customize-control-ast-responsive-color .customize-control-content .ast-color-picker-alpha.' + device ).addClass( 'active' );
-
-		jQuery( '.customize-control-ast-responsive-color .ast-responsive-btns li' ).removeClass( 'active' );
-
-		jQuery( '.customize-control-ast-responsive-color .ast-responsive-btns li.' + device ).addClass( 'active' );
-
-		jQuery('.wp-full-overlay-footer .devices button').on('click', function() {
-
-			var device = jQuery(this).attr('data-device');
-
-			jQuery( '.customize-control-ast-responsive-color .customize-control-content .ast-color-picker-alpha' ).removeClass( 'active' );
-			jQuery( '.customize-control-ast-responsive-color .customize-control-content .ast-responsive-color.' + device ).addClass( 'active' );
-		});
-
-		this.container.find( '.ast-responsive-btns button' ).on( 'click', function( event ) {
-
-			var device = jQuery(this).attr('data-device');
-			if( 'desktop' == device ) {
-				device = 'tablet';
-			} else if( 'tablet' == device ) {
-				device = 'mobile';
-			} else {
-				device = 'desktop';
-			}
-
-			jQuery( '.wp-full-overlay-footer .devices button[data-device="' + device + '"]' ).trigger( 'click' );
-		});
+		getResponsiveColorJs( this );
 	},
 
 } );
