+++ inc/customizer/react/src/responsive/control.js
@@ -1,4 +1,5 @@
 import ResponsiveComponent from './responsive-component.js';
+import { getResponsiveJs } from '../common/responsive-helper';
 
 export const responsiveControl = wp.customize.astraControl.extend( {
 	renderContent: function renderContent() {
@@ -7,38 +8,6 @@ export const responsiveControl = wp.customize.astraControl.extend( {
 	},
 	ready: function() {
 
-		'use strict';
-
-		let device = jQuery('.wp-full-overlay-footer .devices button.active').attr('data-device')
-
-		jQuery( '.customize-control-ast-responsive .input-wrapper input' ).removeClass( 'active' );
-
-		jQuery( '.customize-control-ast-responsive .input-wrapper input.' + device ).addClass( 'active' );
-
-		jQuery( '.customize-control-ast-responsive .ast-responsive-btns li' ).removeClass( 'active' );
-
-		jQuery( '.customize-control-ast-responsive .ast-responsive-btns li.' + device ).addClass( 'active' );
-
-		jQuery('.wp-full-overlay-footer .devices button').on('click', function() {
-
-			var device = jQuery(this).attr('data-device');
-
-			jQuery( '.customize-control-ast-responsive .input-wrapper input, .customize-control .ast-responsive-btns > li' ).removeClass( 'active' );
-			jQuery( '.customize-control-ast-responsive .input-wrapper input.' + device + ', .customize-control .ast-responsive-btns > li.' + device ).addClass( 'active' );
-		});
-
-		this.container.find( '.ast-responsive-btns button' ).on( 'click', function( event ) {
-
-			var device = jQuery(this).attr('data-device');
-			if( 'desktop' == device ) {
-				device = 'tablet';
-			} else if( 'tablet' == device ) {
-				device = 'mobile';
-			} else {
-				device = 'desktop';
-			}
-
-			jQuery( '.wp-full-overlay-footer .devices button[data-device="' + device + '"]' ).trigger( 'click' );
-		});
+		getResponsiveJs( this );
 	},
 } );
